rename: shared 폴더로 types 이동
rename: shared 폴더로 types 이동
